+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/completion/EnumValueNameProvider.kt
@@ -3,6 +3,8 @@ package io.kanro.idea.plugin.protobuf.lang.completion
 import com.intellij.codeInsight.completion.CompletionParameters
 import com.intellij.codeInsight.completion.CompletionProvider
 import com.intellij.codeInsight.completion.CompletionResultSet
+import com.intellij.codeInsight.completion.InsertHandler
+import com.intellij.codeInsight.lookup.LookupElement
 import com.intellij.codeInsight.lookup.LookupElementBuilder
 import com.intellij.psi.util.parentOfType
 import com.intellij.util.ProcessingContext
@@ -21,11 +23,16 @@ object EnumValueNameProvider : CompletionProvider<CompletionParameters>() {
         result.addElement(
             LookupElementBuilder.create("$name unspecified".toScreamingSnakeCase())
                 .withTypeText("enum value")
+                .withInsertHandler(enumValueNumberInserter(0))
         )
         result.addElement(
             LookupElementBuilder.create(name.toScreamingSnakeCase())
                 .withTypeText("enum value")
                 .withInsertHandler(SmartInsertHandler("_"))
         )
+    }
+
+    private fun enumValueNumberInserter(number: Long): InsertHandler<LookupElement> {
+        return SmartInsertHandler(" = $number;", -1)
     }
 }